% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topReviGO.R
\name{csvFilePreparationForTopReviGo}
\alias{csvFilePreparationForTopReviGo}
\title{csvFilePreparationForTopReviGo}
\usage{
csvFilePreparationForTopReviGo(csvFile)
}
\arguments{
\item{csvFile}{The csv file containing the genes, that has to be prepared.}
}
\value{
The genes list
}
\description{
csvFilePreparationForTopReviGo
}
