% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topReviGO.R
\name{topReviGO}
\alias{topReviGO}
\title{Chain topGO and REViGO analyses to produce treemaps}
\usage{
topReviGO(geneList, prefix, mapFile, ontology = "BP", mapOrDb = "map",
  p = 0.01)
}
\arguments{
\item{geneList}{The gene list must be a csv file without column name, each
line consisting of the gene name and a 1 or 0, separated by a ",". The 1 or
0 corresponds to the fact that the gene is respectively selected or not.
What I mean by that is that the gene has previously been recognized by the
user as interesting, like belonging to a cluster, or selected by any other
way. It corresponds to the "Predefined list of interesting genes" from the
topGO vignette
(http://www.bioconductor.org/packages/release/bioc/vignettes/topGO/inst/doc/topGO.pdf).
If you don't have a predefined list, please do the previous steps of the
topGO vignette (before 4.4).}

\item{prefix}{A prefix for the outputs.}

\item{mapFile}{A file containing the correspondance between the gene name and
a GO name, 1 per line, in the format :
GeneName<tabulation>GOName
It can also be a Db name, thus one needs to change the option mapOrDb.}

\item{ontology}{"BP", "CC" or "MF" for Biological Process, Cellular Component
or Molecular Function, this is the GO categories outputted by REViGO.}

\item{mapOrDb}{map if a map file is used, db if a database name is provided.}

\item{p}{the p-value for the weight Fisher test to accept the GO. The Ficher
test based on weights takes multiple testing into account directly without
further need to apply a fdr or bonferroni correction.}
}
\value{
A csv file containing the enriched GO terms and a treemap pdf file
containing the image.
}
\description{
This package takes a list of genes, a map file with the
correspondance between gene name and GO annotation and a prefix for the
output. It will then do a topGO analysis and send the results to the
REViGO (http://revigo.irb.hr/) website, to summarize the list of GO and
produce a treemap.
}
\details{
The goal is from a list of genes and the corresponding GO map, to
be able to produce an enriched list of GO annotations and a treemap to
easily visualize. By default, the biological process is outputted, but
it can also be the Cellular Component or the Molecular Function. One can use
a installed db or a map file.
}
\examples{
library(hgu133a.db)
selGenes <- sample(ls(hgu133aGO), 50)
allGenes <-  factor(as.integer(ls(hgu133aGO) \%in\% selGenes))
names(allGenes) <- ls(hgu133aGO)
topReviGO(allGenes, "toto", "hgu133a", mapOrDb = "db", p = 0.01)
}
